test: fold torch._scaled_mm comparison into 4wave test
Drop the separate bench_fp8_gemm_4wave_compare.py and surface the
same comparison via a --vs_torch flag on tests/kernels/test_fp8_gemm_4wave.py
so input quantization and perf harness are identical.
